Add return-void-barrier instruction.
This introduces the return-void-barrier instruction, which is identical
to return-void on UP systems, but provides an additional store/store
barrier on SMP.  This is intended for use in constructors of objects
with final fields.

The assembler doesn't like "dmb st", and we don't have an
ANDROID_MEMBAR_STORE barrier defined, so this currently uses full
fences.

This just defines the new instruction.  It's not actually used yet.

Also, removed some stale "unused" files from the x86 and x86-atom
directories.

Bug 2965743.

Change-Id: I072e372fd2d57f2617a8d4fff5fd4b38bdda75d1

Conflicts:
	libdex/DexOpcodes.h
	libdex/InstrUtils.c
	vm/mterp/config-x86-atom
	vm/mterp/out/InterpC-x86-atom.cpp