Added a barrier to final field updates.
When deserializing an object, code can write to final fields.  The JMM
requires that we give final fields written this way the same treatment
as we would final fields written directly in a constructor.

This adds memory barriers to the two reflection field-set handlers
when the field in question is marked "final".

Also, added a final field update to test 046.

Bug 2965743.

Change-Id: I5ea3cb49e04d932a655740058b847dcb8712e041

Conflicts:
	vm/native/java_lang_reflect_Field.cpp