Switch to VMWAIT before sending to DDMS
If you stop DDMS with Ctrl-Z, packets sent from the VM will start
to back up and eventually the network write will block.  The VM was
doing the write in "running" state, which meant that if another thread
tried to GC it would be unable to suspend the stalled thread and the
VM would abort.

Now when this happens the affected thread(s) will be quietly stuck until
DDMS is resumed.  (We may want to fix that eventually.)

Bug 3101425.

Change-Id: Ide83c4622137e24c23ee85c006b497d84cf5006b