Separate HPROF from the GC.
In the beginning, the only way to traverse the roots and heap was to
piggyback off the garbage collector.  As such, HPROF was implemented
by instrumenting the root- and object traversal routines to check a
mode flag and call into HPROF during a GC when the flag was set.

This change moves the HPROF calls out of the GC and into callbacks
invoked through the visitor.  Notably, it allows HPROF dumps to be
computed at any point in time without invoking a GC and potentially
destroying evidence relating to the cause of an OOM.

Change-Id: I2b74c4f10f35af3ca33b7c0bbfe470a8b586ff66

Conflicts:
	vm/alloc/GC.h
	vm/alloc/Heap.cpp
	vm/alloc/MarkSweep.cpp
	vm/alloc/Verify.cpp
	vm/alloc/Visit.cpp
	vm/alloc/Visit.h