Fixed identification of objects in the Constant Collector.
The constant collector was improperly identifying constant objects,
keying off the move-result-pseudo instruction that gets the result
instead of the const-string/class instruction that generates the
result. The new code fixes this, so that the following block of code
detects that the const-string/class instruction can throw and will
properly ignore objects. Before, since move-result-pseudo can't throw,
it would skip past that block and try to collect the object anyway.

Change-Id: Ia0a8b07cbe5ece8f586581b86e8507d29a2b9e02